chore: move cursor bubble to internal
