+++ internal/cursor/cursor.go
